refact: new composition method by page
