+++ src/pages/Home/Home.tsx
@@ -1,71 +1,11 @@
 import React, { useContext, useEffect } from 'react';
-import { Variants, motion } from 'framer-motion';
-import { EASE_SWEET } from '@constants/animations';
+import { motion } from 'framer-motion';
 import { ThemeContext } from '@context/ThemeProvider';
 import badgeLight from '@assets/images/badge_light.png';
 import badgeDark from '@assets/images/badge_dark.png';
-import Signature from '@components/Signature';
 import MouseFollower from '@components/MouseFollower';
+import Signature from './components/Signature';
 import GoghFrame from './components/GoghFrame/Composition';
-
-const letterVariants: Variants = {
-  initial: {
-    opacity: 0,
-    y: 100,
-  },
-  animate: {
-    opacity: 1,
-    y: 0,
-    transition: {
-      duration: 0.8,
-      ease: EASE_SWEET,
-    },
-  },
-  exit: {
-    opacity: 0,
-    y: 100,
-    transition: {
-      duration: 0.8,
-      ease: EASE_SWEET,
-    },
-  },
-};
-
-const firstNameLetters = ['R', 'a', 'f', 'a', 'e', 'l'];
-const firstNameVariants: Variants = {
-  animate: {
-    transition: {
-      delayChildren: 1,
-      staggerChildren: 0.04,
-      staggerDirection: -1,
-    },
-  },
-  exit: {
-    transition: {
-      delayChildren: 0,
-      staggerChildren: 0.04,
-      staggerDirection: -1,
-    },
-  },
-};
-
-const lastNameLetters = ['F', 'e', 'r', 'n', 'a', 'n', 'd', 'e', 's'];
-const lastNameVariants: Variants = {
-  animate: {
-    transition: {
-      delayChildren: 1,
-      staggerChildren: 0.04,
-      staggerDirection: 1,
-    },
-  },
-  exit: {
-    transition: {
-      delayChildren: 0,
-      staggerChildren: 0.04,
-      staggerDirection: 1,
-    },
-  },
-};
 
 export default function Home() {
   const { isDarkMode } = useContext(ThemeContext);
@@ -108,22 +48,7 @@ export default function Home() {
 
       <GoghFrame />
 
-      <Signature.Root>
-        <Signature.Text
-          props={{
-            letters: firstNameLetters,
-            phraseVariants: firstNameVariants,
-            letterVariants,
-          }}
-        />
-        <Signature.Text
-          props={{
-            letters: lastNameLetters,
-            phraseVariants: lastNameVariants,
-            letterVariants,
-          }}
-        />
-      </Signature.Root>
+      <Signature />
     </main>
   );
 }
