+++ src/pages/Home/components/Signature/index.ts
@@ -1,0 +1,3 @@
+import Composition from './Composition';
+
+export default Composition;
